// frontend/src/gen/msw-api-ts/models/ApiForecastLineEntry.ts
// tslint:disable
/**
 * Msw Api
 * Msw Api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface ApiForecastLineEntry
 */
export interface ApiForecastLineEntry {
    /**
     * @type {string}
     * @memberof ApiForecastLineEntry
     */
    timestamp?: string;
    /**
     * @type {number}
     * @memberof ApiForecastLineEntry
     */
    flow?: number;
}